lib: librsu: allow copying to file for zero priority slots
Nothing is technically limiting the contents of a zero priority slot
from being copied to a file. For example, in a QSPI layout with 6 RSU
slots, where 2 are dedicated for storing FPGA SOFs and the other 4 are
dedicated for storing U-Boot (SSBL) and Linux/OS images, only the SOF
slots would have non-zero priorities, thus it would not possible using
librsu to dump the content of the non-SOF slots. Removing the priority
check, it is now possible to dump the content of other slots.

Signed-off-by: Jared Baur <jaredbaur@fastmail.com>